almost there w/filtering prs by repo only
