+++ shared/ui/Stream/CrossPostIssueControls/IssueDropdown.tsx
@@ -40,8 +40,7 @@ import { useDidMount } from "@codestream/webview/utilities/hooks";
 import { Modal } from "../Modal";
 import { Button } from "@codestream/webview/src/components/Button";
 import { OpenUrlRequestType, WebviewPanels } from "@codestream/protocols/webview";
-import { Card } from "@codestream/webview/src/components/Card";
-import { ButtonRow } from "../StatusPanel";
+import { ButtonRow } from "@codestream/webview/src/components/Dialog";
 import { Dialog } from "@codestream/webview/src/components/Dialog";
 
 interface ProviderInfo {
