+++ shared/ui/store/preferences/reducer.ts
@@ -4,6 +4,7 @@ import { PreferencesActionsType, PreferencesState, FilterQuery } from "./types";
 import { merge } from "lodash-es";
 import { createSelector } from "reselect";
 import { CodeStreamState } from "..";
+import { PullRequestQuery } from "@codestream/protocols/api";
 
 type PreferencesActions = ActionType<typeof actions>;
 
@@ -29,5 +30,41 @@ export const getSavedSearchFilters = createSelector(
 			savedSearchFilters[parseInt(key, 10)] = preferences.savedSearchFilters[key];
 		});
 		return savedSearchFilters.filter(filter => filter.label.length > 0);
+	}
+);
+
+// FIXME hard-coded github*com
+const DEFAULT_QUERIES: { [index: number]: PullRequestQuery } = {
+	// { name: "Local PR Branches", query: `is:pr author:@me`, hidden: false, repoOnly: true },
+	0: {
+		providerId: "github*com",
+		name: "Waiting on my Review",
+		query: `is:pr is:open review-requested:@me`,
+		hidden: false
+	},
+	1: {
+		providerId: "github*com",
+		name: "Assigned to Me",
+		query: `is:pr is:open assignee:@me`,
+		hidden: false
+	},
+	2: {
+		providerId: "github*com",
+		name: "Created by Me",
+		query: `is:pr is:open author:@me`,
+		hidden: false
+	}
+};
+
+export const getSavedPullRequestQueries = createSelector(
+	(state: CodeStreamState) => state.preferences,
+	(_, providerId: string) => providerId,
+	(preferences, providerId) => {
+		const pullRequestQueries: PullRequestQuery[] = [];
+		const queries = preferences.pullRequestQueries || DEFAULT_QUERIES;
+		Object.keys(queries).forEach(key => {
+			pullRequestQueries[parseInt(key, 10)] = queries[key];
+		});
+		return pullRequestQueries.filter(q => q && q.providerId === providerId && q.query.length > 0);
 	}
 );
