+++ shared/ui/store/providerPullRequests/actions.ts
@@ -143,6 +143,7 @@ export const removeFromMyPullRequests = (providerId: string, id: string) =>
 export const getMyPullRequests = (
 	providerId: string,
 	queries: string[],
+	openReposOnly: boolean,
 	options?: { force?: boolean },
 	throwOnError?: boolean
 ) => async (dispatch, getState: () => CodeStreamState) => {
@@ -170,8 +171,7 @@ export const getMyPullRequests = (
 			providerId: providerId,
 			params: {
 				queries,
-				force: force || (options && options.force),
-				isOpen: true
+				force: force || (options && options.force)
 			}
 		});
